requirements.txt
Pillow==8.0.4
pandas==1.2.3